refactor(language): move recipe API out of configs
