+++ docs/source/language_and_recipes.mdx
@@ -66,7 +66,8 @@ The language stack itself has three internal modules backing layer 1:
 
 ## Layer 2 — recipe anatomy
 
-Recipes are YAML files backed by `TrainingRecipe` and `MessageTurn`. They
+Recipes are YAML files backed by `TrainingRecipe` and `MessageTurn` from
+`lerobot.language.recipe`. They
 declare which annotation rows to pull (via `bindings`) and how to compose them
 into chat turns (`messages`).
