+++ src/lerobot/language/recipe.py
@@ -14,6 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""Recipe definitions, validation, loading and shared message rendering."""
+
 from __future__ import annotations
 
 import copy
